+++ src/Sentry/Internal/Http/HttpTransport.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
@@ -17,7 +19,11 @@ namespace Sentry.Internal.Http
         private readonly HttpClient _httpClient;
         private readonly Action<HttpRequestHeaders> _addAuth;
 
-        internal const string NoMessageFallback = "No message";
+        // Keep track of rate limits and their expiry dates
+        private readonly Dictionary<RateLimitCategory, DateTimeOffset> _categoryLimitResets =
+            new Dictionary<RateLimitCategory, DateTimeOffset>();
+
+        internal const string DefaultErrorMessage = "No message";
 
         public HttpTransport(
             SentryOptions options,
@@ -29,28 +35,96 @@ namespace Sentry.Internal.Http
             _addAuth = addAuth;
         }
 
+        private Envelope ApplyRateLimitsOnEnvelope(Envelope envelope, DateTimeOffset instant)
+        {
+            // Re-package envelope, discarding items that don't fit the rate limit
+            var envelopeItems = new List<EnvelopeItem>();
+            foreach (var envelopeItem in envelope.Items)
+            {
+                // Check if there is at least one matching category for this item that is rate-limited
+                var isRateLimited = _categoryLimitResets
+                    .Where(kvp => kvp.Value > instant)
+                    .Any(kvp => kvp.Key.Matches(envelopeItem));
+
+                if (!isRateLimited)
+                {
+                    envelopeItems.Add(envelopeItem);
+                }
+                else
+                {
+                    _options.DiagnosticLogger?.LogDebug(
+                        "Envelope item of type {0} was discarded because it's rate-limited.",
+                        envelopeItem.TryGetType()
+                    );
+                }
+            }
+
+            return new Envelope(envelope.Header, envelopeItems);
+        }
+
+        private void ExtractRateLimits(HttpResponseMessage response, DateTimeOffset instant)
+        {
+            if (!response.Headers.TryGetValues("X-Sentry-Rate-Limits", out var rateLimitHeaderValues))
+            {
+                return;
+            }
+
+            // Join to a string to handle both single-header and multi-header cases
+            var rateLimitsEncoded = string.Join(",", rateLimitHeaderValues);
+
+            // Parse and order by retry-after so that the longer rate limits are set last (and not overwritten)
+            var rateLimits = RateLimit.ParseMany(rateLimitsEncoded).OrderBy(rl => rl.RetryAfter);
+
+            // Persist rate limits
+            foreach (var rateLimit in rateLimits)
+            {
+                foreach (var rateLimitCategory in rateLimit.Categories)
+                {
+                    _categoryLimitResets[rateLimitCategory] = instant + rateLimit.RetryAfter;
+                }
+            }
+        }
+
         public async Task SendEnvelopeAsync(Envelope envelope, CancellationToken cancellationToken = default)
         {
-            var request = CreateRequest(envelope);
-            var response = await _httpClient.SendAsync(request, cancellationToken).ConfigureAwait(false);
+            var instant = DateTimeOffset.Now;
+
+            // Apply rate limiting and re-package envelope items
+            var processedEnvelope = ApplyRateLimitsOnEnvelope(envelope, instant);
+            if (!processedEnvelope.Items.Any())
+            {
+                _options.DiagnosticLogger?.LogInfo(
+                    "Envelope {0} was discarded because all contained items are rate-limited.",
+                    envelope.TryGetEventId()
+                );
+
+                return;
+            }
+
+            // Send envelope to ingress
+            using var request = CreateRequest(processedEnvelope);
+            using var response = await _httpClient.SendAsync(request, cancellationToken).ConfigureAwait(false);
+
+            // Read & set rate limits for future requests
+            ExtractRateLimits(response, instant);
 
             if (response.StatusCode == HttpStatusCode.OK)
             {
                 _options.DiagnosticLogger?.LogDebug(
                     "Envelope {0} successfully received by Sentry.",
-                    envelope.TryGetEventId()
+                    processedEnvelope.TryGetEventId()
                 );
             }
             else if (_options.DiagnosticLogger?.IsEnabled(SentryLevel.Error) == true)
             {
                 var responseJson = await response.Content.ReadAsJsonAsync().ConfigureAwait(false);
-                var errorMessage = responseJson.SelectToken("detail")?.Value<string>() ?? NoMessageFallback;
+                var errorMessage = responseJson.SelectToken("detail")?.Value<string>() ?? DefaultErrorMessage;
 
                 _options.DiagnosticLogger?.Log(
                     SentryLevel.Error,
                     "Sentry rejected the envelope {0}. Status code: {1}. Sentry response: {2}",
                     null,
-                    envelope.TryGetEventId(),
+                    processedEnvelope.TryGetEventId(),
                     response.StatusCode,
                     errorMessage
                 );